requirements.txt
fastapi~=0.70.0
loguru~=0.6.0
pydantic~=1.8.2
boto3~=1.15.18
botocore~=1.18.18
PyYAML~=5.4.1